requirements.txt
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
deepdiff==8.0.1
fonttools==4.54.1
gurobipy==11.0.3
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
networkx==3.2.1
numpy==1.26.4
orderly-set==5.2.2
packaging==24.1
pandapower==2.14.11
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scipy==1.13.1
six==1.16.0
tqdm==4.66.5
tzdata==2024.2
zipp==3.20.2
